requirements.txt
faiss_cpu==1.10.0
Requests==2.32.3
sentence_transformers==3.4.1
unstructured==0.16.23
